requirements_streamlit.txt
streamlit==1.39.0
tensorflow==2.20.0
numpy==2.1.3
pillow==11.2.1
